removed notes.svg
