+++ lib/formatPoints.tsx
@@ -1,0 +1,3 @@
+export function formatPoints(points: any) {
+  return `${(points / 1000).toFixed(1)}K`;
+}
